+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 * errors return status of 2 instead of 1
+* capture invalid option error, and display message
 
 ### New features